requirements.txt
matplotlib>=3.7.5
numpy>=1.24.1
ogb>=1.3.6
PyYAML>=6.0.1
rich>=14.0.0
scikit_learn>=1.3.2
scipy>=1.9.1
seaborn>=0.13.2
setuptools>=68.2.2
torch>=2.1.0
torch_cluster>=1.6.2
torch_geometric>=2.5.1
torch_sparse>=0.6.18
umap_learn>=0.5.7
yacs>=0.1.8
